+++ review/go.mod
@@ -8,11 +8,11 @@ replace goondex/index v0.0.0 => ./pkg/index
 
 replace goondex/engine v0.0.0 => ./pkg/engine
 
-replace goondex/goondex v0.0.0 => ./pkg/goondex
+replace goondex/web v0.0.0 => ./pkg/web
 
 require (
 	goondex/crawler v0.0.0 // indirect
 	goondex/engine v0.0.0 // indirect
-	goondex/goondex v0.0.0 // indirect
+	goondex/web v0.0.0 // indirect
 	goondex/index v0.0.0 // indirect
 )
