--- review/pkg/goondex/go.mod
@@ -1,3 +1,0 @@
-module goondex
-
-go 1.15
