+++ review/pkg/web/go.mod
@@ -1,0 +1,3 @@
+module web
+
+go 1.15
